fix(frontend): 临时密码修改调整及优化 #8250 # Reviewed, transaction id: 25459
